Gather the log files before the actual outputs and mark the log files gatherers as intermediates. Since the outputs will only be gathered iff the logs were gathered this allows the job name to change without causing SG to re-run.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@5285 348d0f76-0448-11de-a6fe-93d51630548a

+++ scala/src/org/broadinstitute/sting/queue/engine/QGraph.scala
@@ -484,7 +484,7 @@ class QGraph extends Logging {
   }
 
   private def failedDescription(failed: FunctionEdge) = {
-    var description = new StringBuilder
+    val description = new StringBuilder
     if (settings.retries > 0)
       description.append("Attempt %d of %d.%n".format(failed.retries + 1, settings.retries + 1))
     description.append(failed.function.description)
@@ -749,7 +749,7 @@ class QGraph extends Logging {
     if (running && !settings.keepIntermediates && success) {
       logger.info("Deleting intermediate files.")
       traverseFunctions(edge => {
-        if (edge.function.isIntermediate) {
+        if (edge.function.isIntermediate && edge.function.deleteIntermediateOutputs) {
           logger.debug("Deleting intermediates:" + edge.function.description)
           edge.function.deleteOutputs()
         }
